refactor: extract environment/analytics helpers from store into utils
store.tsx carried IAS environment/analytics utilities (getInitialEnvironment,
IASEnvironment, and their private helpers) that don't belong in the state
store. Move them to a new app/src/utils/environment.ts, and move the Mode
enum to app/src/constants.ts alongside it. Retarget every consumer to the
new modules with a clean cut (no compatibility re-exports left in store.tsx).

Pure refactor, no behavior change. Affects both bcwallet and bcsc build
targets since app/src is shared.

Refs #3583
