+++ app/src/bcsc-theme/features/fcm/FcmViewModel.test.ts
@@ -1,9 +1,9 @@
+import { Mode } from '@/constants'
 import { AppError } from '@/errors/appError'
 import { AppEventCode } from '@/events/appEventCode'
 import { DeviceEventEmitter } from 'react-native'
 import { decodeLoginChallenge, showLocalNotification } from 'react-native-bcsc-core'
 import { BCSCEventTypes } from '../../../events/eventTypes'
-import { Mode } from '../../../store'
 import { getBCSCApiClient } from '../../contexts/BCSCApiClientContext'
 import { BCSCEvent, BCSCReason } from '../../utils/id-token'
 import { PairingService } from '../pairing'
