+++ app/src/bcsc-theme/features/fcm/FcmViewModel.ts
@@ -1,3 +1,4 @@
+import { Mode } from '@/constants'
 import { AppError } from '@/errors/appError'
 import { ErrorRegistry } from '@/errors/errorRegistry'
 import { toAppError } from '@bcsc-theme/utils/native-error-map'
@@ -6,7 +7,6 @@ import { getApp } from '@react-native-firebase/app'
 import { getMessaging, getToken } from '@react-native-firebase/messaging'
 import { Platform } from 'react-native'
 import { decodeLoginChallenge, JWK, showLocalNotification } from 'react-native-bcsc-core'
-import { Mode } from '../../../store'
 import { getBCSCApiClient } from '../../contexts/BCSCApiClientContext'
 import { isVerificationRequestReviewed } from '../../utils/id-token'
 import { PairingService } from '../pairing'
