+++ app/src/bcsc-theme/features/services/Services.tsx
@@ -5,8 +5,9 @@ import useDataLoader from '@/bcsc-theme/hooks/useDataLoader'
 import { useTokenService } from '@/bcsc-theme/services/hooks/useTokenService'
 import { BCSCMainStackParams, BCSCScreens } from '@/bcsc-theme/types/navigators'
 import { getCardProcessForCardType } from '@/bcsc-theme/utils/card-utils'
+import { Mode } from '@/constants'
 import { useDebounce } from '@/hooks/useDebounce'
-import { BCState, Mode } from '@/store'
+import { BCState } from '@/store'
 import { testIdWithKey, ThemedText, TOKENS, useServices, useStore, useTheme } from '@bifold/core'
 import { useFocusEffect, useNavigation } from '@react-navigation/native'
 import { StackNavigationProp } from '@react-navigation/stack'
