+++ app/src/screens/IASEnvironment.tsx
@@ -1,9 +1,10 @@
 import { useFactoryReset } from '@/bcsc-theme/api/hooks/useFactoryReset'
 import { Analytics } from '@/utils/analytics/analytics-singleton'
 import { TOKENS, useServices, useStore } from '@bifold/core'
+import { IASEnvironment } from '@utils/environment'
 import React, { useCallback } from 'react'
 import { setIssuer } from 'react-native-bcsc-core'
-import { BCDispatchAction, BCState, IASEnvironment } from '../store'
+import { BCDispatchAction, BCState } from '../store'
 import EnvironmentSelector from './EnvironmentSelector'
 
 interface IASEnvironmentProps {
